requirements.txt
requests
prometheus_client
python-telegram-bot
urllib3==1.26.7
python_dotenv
